backend/requirements_production.txt
# Core Framework - Latest compatible versions
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.2

# ML & Data Science
scikit-learn==1.5.2
numpy==1.26.4
tensorflow==2.16.1
pandas==2.2.2

# System & Monitoring - Updated for safety compatibility
psutil==6.1.0
websockets==12.0
python-dotenv==1.0.1
requests==2.32.3
aiofiles==24.1.0

# Security & Networking
scapy==2.5.0
cryptography==42.0.8
PyJWT==2.9.0

# Federated Learning
flwr==1.11.0

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Additional Tools - Fixed compatibility
jinja2==3.1.4
python-multipart==0.0.12
rich==13.8.1
click==8.1.7

# Visualization & Charts
plotly==5.22.0
dash==2.17.1
matplotlib==3.9.2
seaborn==0.13.2
bokeh==3.5.2 


libcap
NPcap